MASSIVE REFACTOR OF INTERFACE LAYER
