+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/wfReactInterface/interface.ts
@@ -1,63 +1,31 @@
-import _ from 'lodash';
-import LRUCache from 'lru-cache';
-import {useEffect, useMemo, useState} from 'react';
+import {OBJECT_CATEGORIES, OP_CATEGORIES} from './constants';
 
-import {
-  constBoolean,
-  constFunction,
-  constString,
-  Node,
-  opAnd,
-  opArray,
-  opArtifactLastMembership,
-  opArtifactMembershipArtifactVersion,
-  opArtifactName,
-  opArtifactTypeArtifacts,
-  opArtifactVersionArtifactSequence,
-  opArtifactVersionCreatedAt,
-  opArtifactVersionHash,
-  opArtifactVersionIsWeaveObject,
-  opArtifactVersionMetadata,
-  opArtifactVersions,
-  opArtifactVersionVersionId,
-  opDict,
-  opFilter,
-  opFlatten,
-  opIsNone,
-  opJoin,
-  opMap,
-  opPick,
-  opProjectArtifact,
-  opProjectArtifactTypes,
-  opProjectArtifactVersion,
-  opRootProject,
-  opStringEqual,
-  typedDict,
-} from '../../../../../../core';
-import {useDeepMemo} from '../../../../../../hookUtils';
-import {useNodeValue} from '../../../../../../react';
-import {Span, SpanWithFeedback} from '../../../Browse2/callTree';
-import {fnRunsNode} from '../../../Browse2/callTreeHooks';
-import {PROJECT_CALL_STREAM_NAME, WANDB_ARTIFACT_REF_PREFIX} from './constants';
-import * as trace_server_client from './trace_server_client';
-
-export const OP_CATEGORIES = [
-  'train',
-  'predict',
-  'score',
-  'evaluate',
-  'tune',
-] as const;
-const OBJECT_CATEGORIES = ['model', 'dataset'] as const;
-type OpCategory = (typeof OP_CATEGORIES)[number];
+export type OpCategory = (typeof OP_CATEGORIES)[number];
 export type ObjectCategory = (typeof OBJECT_CATEGORIES)[number];
 
-type Loadable<T> = {
+export interface RawSpanFromStreamTableEra {
+  name: string;
+  inputs: {_keys?: string[]; [key: string]: any};
+  output: undefined | {_keys?: string[]; [key: string]: any};
+  status_code: string; // TODO enum
+  exception?: string;
+  attributes: {[key: string]: any};
+  summary: {latency_s: number; [key: string]: any};
+  span_id: string;
+  trace_id: string;
+  parent_id?: string;
+  timestamp: number;
+  start_time_ms: number;
+  end_time_ms?: number;
+}
+
+export type RawSpanFromStreamTableEraWithFeedback =
+  RawSpanFromStreamTableEra & {feedback?: any};
+
+export type Loadable<T> = {
   loading: boolean;
   result: T | null;
 };
-
-type RefUri = string;
 
 export type CallKey = {
   entity: string;
@@ -70,166 +38,8 @@ export type CallSchema = CallKey & {
   opVersionRef: string | null;
   traceId: string;
   parentId: string | null;
-  rawSpan: Span;
+  rawSpan: RawSpanFromStreamTableEra;
   rawFeedback?: any;
-};
-
-const convertISOToDate = (iso: string) => {
-  return new Date(iso);
-};
-
-const traceCallToLegacySpan = (
-  traceCall: trace_server_client.TraceCallSchema
-): Span => {
-  const startDate = convertISOToDate(traceCall.start_datetime);
-  const endDate = traceCall.end_datetime
-    ? convertISOToDate(traceCall.end_datetime)
-    : null;
-  let statusCode = 'UNSET';
-  if (traceCall.exception) {
-    statusCode = 'ERROR';
-  } else if (traceCall.end_datetime) {
-    statusCode = 'SUCCESS';
-  }
-  let latencyS = 0;
-  if (startDate && endDate) {
-    latencyS = (endDate.getTime() - startDate.getTime()) / 1000;
-  }
-  const summary = {
-    latency_s: latencyS,
-    ...(traceCall.summary ?? {}),
-  };
-  return {
-    name: traceCall.name,
-    inputs: traceCall.inputs,
-    output: traceCall.outputs,
-    status_code: statusCode,
-    exception: traceCall.exception,
-    attributes: traceCall.attributes,
-    summary,
-    span_id: traceCall.id,
-    trace_id: traceCall.trace_id,
-    parent_id: traceCall.parent_id,
-    timestamp: startDate.getTime(),
-    start_time_ms: startDate.getTime(),
-    end_time_ms: endDate?.getTime(),
-  };
-};
-
-const traceCallToUICallSchema = (
-  traceCall: trace_server_client.TraceCallSchema
-): CallSchema => {
-  return {
-    entity: traceCall.entity,
-    project: traceCall.project,
-    callId: traceCall.id,
-    traceId: traceCall.trace_id,
-    parentId: traceCall.parent_id ?? null,
-    spanName: traceCall.name.startsWith(WANDB_ARTIFACT_REF_PREFIX)
-      ? refUriToOpVersionKey(traceCall.name).opId
-      : traceCall.name,
-    opVersionRef: traceCall.name.startsWith(WANDB_ARTIFACT_REF_PREFIX)
-      ? traceCall.name
-      : null,
-    rawSpan: traceCallToLegacySpan(traceCall),
-    rawFeedback: {},
-  };
-};
-
-export const spanToCallSchema = (
-  entity: string,
-  project: string,
-  span: SpanWithFeedback
-): CallSchema => {
-  // This rawSpan construction fixes issues with crashed runs using the
-  // streamtable graph client (will be fixed in the future)
-  const rawSpan = span as Span;
-  rawSpan.summary = span.summary ?? {
-    latency_s: 0,
-  };
-  rawSpan.summary.latency_s = span.summary?.latency_s ?? 0;
-  rawSpan.status_code = span.status_code ?? 'UNSET';
-  rawSpan.inputs = span.inputs ?? {};
-
-  return {
-    entity,
-    project,
-    callId: span.span_id,
-    traceId: span.trace_id,
-    parentId: span.parent_id ?? null,
-    spanName: span.name.startsWith(WANDB_ARTIFACT_REF_PREFIX)
-      ? refUriToOpVersionKey(span.name).opId
-      : span.name,
-    opVersionRef: span.name.startsWith(WANDB_ARTIFACT_REF_PREFIX)
-      ? span.name
-      : null,
-    rawSpan,
-    rawFeedback: span.feedback,
-  };
-};
-
-export const useCall = (key: CallKey | null): Loadable<CallSchema | null> => {
-  const cachedCall = key ? getCallFromCache(key) : null;
-  const [callRes, setCallRes] =
-    useState<trace_server_client.TraceCallReadRes | null>(null);
-  const deepKey = useDeepMemo(key);
-  useEffect(() => {
-    if (deepKey) {
-      trace_server_client
-        .callRead({
-          entity: deepKey.entity,
-          project: deepKey.project,
-          id: deepKey.callId,
-        })
-        .then(res => {
-          setCallRes(res);
-        });
-    }
-  }, [deepKey]);
-
-  return useMemo(() => {
-    if (key == null) {
-      return {
-        loading: false,
-        result: null,
-      };
-    }
-    if (cachedCall != null) {
-      return {
-        loading: false,
-        result: cachedCall,
-      };
-    }
-    const result = callRes ? traceCallToUICallSchema(callRes.call) : null;
-    if (callRes == null) {
-      return {
-        loading: true,
-        result,
-      };
-    } else {
-      if (result) {
-        setCallInCache(key, result);
-      }
-      return {
-        loading: false,
-        result,
-      };
-    }
-  }, [cachedCall, callRes, key]);
-};
-
-export const useParentCall = (
-  call: CallSchema | null
-): Loadable<CallSchema | null> => {
-  let parentCall = null;
-  if (call && call.parentId) {
-    parentCall = {
-      entity: call.entity,
-      project: call.project,
-      callId: call.parentId,
-    };
-  }
-  return useCall(parentCall);
 };
 
 export type CallFilter = {
@@ -245,131 +55,11 @@ export type CallFilter = {
   opCategory?: OpCategory[];
 };
 
-const callsNode = (
-  entity: string,
-  project: string,
-  filter: CallFilter
-): Node => {
-  return fnRunsNode(
-    {
-      entityName: entity,
-      projectName: project,
-      streamName: PROJECT_CALL_STREAM_NAME,
-    },
-    {
-      opUris: filter.opVersionRefs,
-      inputUris: filter.inputObjectVersionRefs,
-      outputUris: filter.outputObjectVersionRefs,
-      traceId: filter.traceId,
-      parentIds: filter.parentIds,
-      traceRootsOnly: filter.traceRootsOnly,
-      callIds: filter.callIds,
-    }
-  );
-};
-export const useCalls = (
-  entity: string,
-  project: string,
-  filter: CallFilter,
-  limit?: number,
-  opts?: {skip?: boolean}
-): Loadable<CallSchema[]> => {
-  const [callRes, setCallRes] =
-    useState<trace_server_client.TraceCallsQueryRes | null>(null);
-  const deepFilter = useDeepMemo(filter);
-  useEffect(() => {
-    if (opts?.skip) {
-      return;
-    }
-    trace_server_client
-      .callsQuery({
-        entity,
-        project,
-        filter: {
-          op_version_refs: deepFilter.opVersionRefs,
-          input_object_version_refs: deepFilter.inputObjectVersionRefs,
-          output_object_version_refs: deepFilter.outputObjectVersionRefs,
-          parent_ids: deepFilter.parentIds,
-          trace_ids: deepFilter.traceId ? [deepFilter.traceId] : undefined,
-          call_ids: deepFilter.callIds,
-          trace_roots_only: deepFilter.traceRootsOnly,
-        },
-        limit,
-      })
-      .then(res => {
-        setCallRes(res);
-      });
-  }, [entity, project, deepFilter, limit, opts?.skip]);
-
-  return useMemo(() => {
-    if (opts?.skip) {
-      return {
-        loading: false,
-        result: [],
-      };
-    }
-    const allResults = (callRes?.calls ?? []).map(traceCallToUICallSchema);
-    const result = allResults.filter((row: any) => {
-      return (
-        deepFilter.opCategory == null ||
-        (row.opVersionRef &&
-          deepFilter.opCategory.includes(
-            opVersionRefOpCategory(row.opVersionRef) as OpCategory
-          ))
-      );
-    });
-
-    if (callRes == null) {
-      return {
-        loading: true,
-        result,
-      };
-    } else {
-      allResults.forEach(call => {
-        setCallInCache(
-          {
-            entity,
-            project,
-            callId: call.callId,
-          },
-          call
-        );
-      });
-      return {
-        loading: false,
-        result,
-      };
-    }
-  }, [callRes, deepFilter.opCategory, entity, project, opts?.skip]);
-};
-
-type OpVersionKey = {
+export type OpVersionKey = {
   entity: string;
   project: string;
   opId: string;
   versionHash: string;
-};
-
-export const refUriToOpVersionKey = (refUri: RefUri): OpVersionKey => {
-  const refDict = refStringToRefDict(refUri);
-  if (
-    refDict.filePathParts.length !== 1 ||
-    refDict.refExtraTuples.length !== 0 ||
-    refDict.filePathParts[0] !== 'obj'
-  ) {
-    if (refDict.versionCommitHash !== '*') {
-      throw new Error('Invalid refUri: ' + refUri);
-    }
-  }
-  return {
-    entity: refDict.entity,
-    project: refDict.project,
-    opId: refDict.artifactName,
-    versionHash: refDict.versionCommitHash,
-  };
-};
-export const opVersionKeyToRefUri = (key: OpVersionKey): RefUri => {
-  return `${WANDB_ARTIFACT_REF_PREFIX}${key.entity}/${key.project}/${key.opId}:${key.versionHash}/obj`;
 };
 
 export type OpVersionSchema = OpVersionKey & {
@@ -379,268 +69,20 @@ export type OpVersionSchema = OpVersionKey & {
   category: OpCategory | null;
 };
 
-const artifactVersionNodeToOpVersionDictNode = (
-  artifactVersionNode: Node<'artifactVersion'>
-) => {
-  const versionIndexNode = opArtifactVersionVersionId({
-    artifactVersion: artifactVersionNode,
-  });
-  const createdAtNode = opArtifactVersionCreatedAt({
-    artifactVersion: artifactVersionNode,
-  });
-  return opDict({
-    missing: opIsNone({val: artifactVersionNode}),
-    versionIndex: versionIndexNode,
-    createdAtMs: createdAtNode,
-  } as any);
-};
-
-export const useOpVersion = (
-  // Null value skips
-  key: OpVersionKey | null
-): Loadable<OpVersionSchema | null> => {
-  const cachedOpVersion = key ? getOpVersionFromCache(key) : null;
-  const artifactVersionNode = opProjectArtifactVersion({
-    project: opRootProject({
-      entity: constString(key?.entity ?? ''),
-      project: constString(key?.project ?? ''),
-    }),
-    artifactName: constString(key?.opId ?? ''),
-    artifactVersionAlias: constString(key?.versionHash ?? ''),
-  });
-  const dataNode = artifactVersionNodeToOpVersionDictNode(
-    artifactVersionNode as any
-  );
-  const dataValue = useNodeValue(dataNode, {
-    skip: key == null || cachedOpVersion != null,
-  });
-  return useMemo(() => {
-    if (key == null) {
-      return {
-        loading: false,
-        result: null,
-      };
-    } else if (cachedOpVersion != null) {
-      return {
-        loading: false,
-        result: cachedOpVersion,
-      };
-    }
-    const result =
-      dataValue.result == null || dataValue.result.missing
-        ? null
-        : {
-            ...key,
-            versionIndex: dataValue.result.versionIndex as number,
-            category: opNameToCategory(key.opId as string),
-            createdAtMs: dataValue.result.createdAtMs as number,
-          };
-    if (dataValue.loading) {
-      return {
-        loading: true,
-        result,
-      };
-    } else {
-      if (result) {
-        setOpVersionInCache(key, result);
-      }
-      return {
-        loading: false,
-        result,
-      };
-    }
-  }, [cachedOpVersion, dataValue.loading, dataValue.result, key]);
-};
-
-type OpVersionFilter = {
+export type OpVersionFilter = {
   // Filters are ANDed across the fields and ORed within the fields
   category?: OpCategory[];
   opIds?: string[];
   latestOnly?: boolean;
 };
 
-export const useOpVersionsNode = (
-  entity: string,
-  project: string,
-  filter: OpVersionFilter,
-  opts?: {skip?: boolean}
-): Node => {
-  const projectNode = opRootProject({
-    entityName: constString(entity),
-    projectName: constString(project),
-  });
-  let artifactsNode = opArray({} as any);
-
-  if (filter.opIds == null) {
-    artifactsNode = opFlatten({
-      arr: opArtifactTypeArtifacts({
-        artifactType: opProjectArtifactTypes({
-          project: projectNode,
-        }),
-      }) as any,
-    });
-  } else {
-    artifactsNode = opArray(
-      _.fromPairs(
-        filter.opIds.map(opId => {
-          return [
-            opId,
-            opProjectArtifact({
-              project: projectNode,
-              artifactName: constString(opId),
-            }),
-          ];
-        })
-      ) as any
-    );
-  }
-
-  let artifactVersionsNode = opArray({} as any);
-  if (filter.latestOnly) {
-    artifactVersionsNode = opArtifactMembershipArtifactVersion({
-      artifactMembership: opArtifactLastMembership({
-        artifact: artifactsNode,
-      }),
-    }) as any;
-  } else {
-    artifactVersionsNode = opFlatten({
-      arr: opArtifactVersions({
-        artifact: artifactsNode,
-      }) as any,
-    });
-  }
-
-  // Filter to only Weave Objects
-  const weaveObjectsNode = opFilter({
-    arr: artifactVersionsNode,
-    filterFn: constFunction({row: 'artifactVersion'}, ({row}) => {
-      return opAnd({
-        lhs: opArtifactVersionIsWeaveObject({artifactVersion: row}),
-        rhs: opStringEqual({
-          lhs: opPick({
-            obj: opArtifactVersionMetadata({
-              artifactVersion: row,
-            }),
-            key: constString('_weave_meta.type_name'),
-          }),
-          rhs: constString('OpDef'),
-        }),
-      });
-    }),
-  });
-
-  // Build Keys
-  const dataNode = opMap({
-    arr: weaveObjectsNode,
-    mapFn: constFunction({row: 'artifactVersion'}, ({row}) => {
-      return opDict({
-        opId: opArtifactName({
-          artifact: opArtifactVersionArtifactSequence({artifactVersion: row}),
-        }),
-        versionHash: opArtifactVersionHash({artifactVersion: row}),
-        dataDict: artifactVersionNodeToOpVersionDictNode(row as any),
-      } as any);
-    }),
-  });
-
-  return dataNode;
-};
-
-export const useOpVersions = (
-  entity: string,
-  project: string,
-  filter: OpVersionFilter,
-  opts?: {skip?: boolean}
-): Loadable<OpVersionSchema[]> => {
-  const dataNode = useOpVersionsNode(entity, project, filter);
-
-  const dataValue = useNodeValue(dataNode, {skip: opts?.skip});
-
-  return useMemo(() => {
-    if (opts?.skip) {
-      return {
-        loading: false,
-        result: [],
-      };
-    }
-    const result = (dataValue.result ?? [])
-      .map((row: any) => ({
-        entity,
-        project,
-        opId: row.opId as string,
-        versionHash: row.versionHash as string,
-        path: 'obj',
-        refExtra: null,
-        versionIndex: row.dataDict.versionIndex as number,
-        typeName: row.dataDict.typeName as string,
-        category: opNameToCategory(row.opId as string),
-        createdAtMs: row.dataDict.createdAtMs as number,
-      }))
-      .filter((row: any) => {
-        return (
-          filter.category == null || filter.category.includes(row.category)
-        );
-      }) as OpVersionSchema[];
-
-    if (dataValue.loading) {
-      return {
-        loading: true,
-        result,
-      };
-    } else {
-      result.forEach(op => {
-        setOpVersionInCache(
-          {
-            entity,
-            project,
-            opId: op.opId,
-            versionHash: op.versionHash,
-          },
-          op
-        );
-      });
-      return {
-        loading: false,
-        result,
-      };
-    }
-  }, [
-    dataValue.loading,
-    dataValue.result,
-    entity,
-    filter.category,
-    opts?.skip,
-    project,
-  ]);
-};
-
-type ObjectVersionKey = {
+export type ObjectVersionKey = {
   entity: string;
   project: string;
   objectId: string;
   versionHash: string;
   path: string;
   refExtra?: string;
-};
-
-export const refUriToObjectVersionKey = (refUri: RefUri): ObjectVersionKey => {
-  const refDict = refStringToRefDict(refUri);
-  return {
-    entity: refDict.entity,
-    project: refDict.project,
-    objectId: refDict.artifactName,
-    versionHash: refDict.versionCommitHash,
-    path: refDict.filePathParts.join('/'),
-    refExtra: refDict.refExtraTuples
-      .map(t => `${t.edgeType}/${t.edgeName}`)
-      .join('/'),
-  };
-};
-
-export const objectVersionKeyToRefUri = (key: ObjectVersionKey): RefUri => {
-  return `${WANDB_ARTIFACT_REF_PREFIX}${key.entity}/${key.project}/${
-    key.objectId
-  }:${key.versionHash}/${key.path}${key.refExtra ? '#' + key.refExtra : ''}`;
 };
 
 export type ObjectVersionSchema = ObjectVersionKey & {
@@ -651,447 +93,55 @@ export type ObjectVersionSchema = ObjectVersionKey & {
   createdAtMs: number;
 };
 
-const typeNameToCategory = (typeName: string): ObjectCategory | null => {
-  for (const category of OBJECT_CATEGORIES) {
-    if (typeName.toLocaleLowerCase().includes(category)) {
-      return category as ObjectCategory;
-    }
-  }
-  return null;
-};
-
-const opNameToCategory = (opName: string): OpCategory | null => {
-  for (const category of OP_CATEGORIES) {
-    if (opName.toLocaleLowerCase().includes(category)) {
-      return category as OpCategory;
-    }
-  }
-  return null;
-};
-
-const artifactVersionNodeToObjectVersionDictNode = (
-  artifactVersionNode: Node<'artifactVersion'>
-) => {
-  const versionIndexNode = opArtifactVersionVersionId({
-    artifactVersion: artifactVersionNode,
-  });
-  const metadataNode = opArtifactVersionMetadata({
-    artifactVersion: artifactVersionNode,
-  });
-  const typeNameNode = opPick({
-    obj: metadataNode,
-    key: constString('_weave_meta.type_name'),
-  });
-  const createdAtNode = opArtifactVersionCreatedAt({
-    artifactVersion: artifactVersionNode,
-  });
-  return opDict({
-    missing: opIsNone({val: artifactVersionNode}),
-    typeName: typeNameNode,
-    versionIndex: versionIndexNode,
-    createdAtMs: createdAtNode,
-  } as any);
-};
-
-export const useObjectVersion = (
-  // Null value skips
-  key: ObjectVersionKey | null
-): Loadable<ObjectVersionSchema | null> => {
-  const cachedObjectVersion = key ? getObjectVersionFromCache(key) : null;
-  const artifactVersionNode = opProjectArtifactVersion({
-    project: opRootProject({
-      entity: constString(key?.entity ?? ''),
-      project: constString(key?.project ?? ''),
-    }),
-    artifactName: constString(key?.objectId ?? ''),
-    artifactVersionAlias: constString(key?.versionHash ?? ''),
-  });
-  const dataNode = artifactVersionNodeToObjectVersionDictNode(
-    artifactVersionNode as any
-  );
-  const dataValue = useNodeValue(dataNode, {
-    skip: key == null || cachedObjectVersion != null,
-  });
-
-  return useMemo(() => {
-    if (key == null) {
-      return {
-        loading: false,
-        result: null,
-      };
-    }
-    if (cachedObjectVersion != null) {
-      return {
-        loading: false,
-        result: cachedObjectVersion,
-      };
-    }
-    const result =
-      dataValue.result == null || dataValue.result.missing
-        ? null
-        : {
-            ...key,
-            versionIndex: dataValue.result.versionIndex as number,
-            typeName: dataValue.result.typeName as string,
-            category: typeNameToCategory(dataValue.result.typeName as string),
-            createdAtMs: dataValue.result.createdAtMs as number,
-          };
-    if (dataValue.loading) {
-      return {
-        loading: true,
-        result,
-      };
-    } else {
-      if (result) {
-        setObjectVersionInCache(key, result);
-      }
-      return {
-        loading: false,
-        result,
-      };
-    }
-  }, [cachedObjectVersion, dataValue.loading, dataValue.result, key]);
-};
-
-type ObjectVersionFilter = {
+export type ObjectVersionFilter = {
   // Filters are ANDed across the fields and ORed within the fields
   category?: ObjectCategory[];
   objectIds?: string[];
   latestOnly?: boolean;
 };
 
-export const useRootObjectVersionsNode = (
-  entity: string,
-  project: string,
-  filter: ObjectVersionFilter,
-  opts?: {skip?: boolean}
-): Node => {
-  const projectNode = opRootProject({
-    entityName: constString(entity),
-    projectName: constString(project),
-  });
-  let artifactsNode = opArray({} as any);
-
-  if (filter.objectIds == null) {
-    artifactsNode = opFlatten({
-      arr: opArtifactTypeArtifacts({
-        artifactType: opProjectArtifactTypes({
-          project: projectNode,
-        }),
-      }) as any,
-    });
-  } else {
-    artifactsNode = opArray(
-      _.fromPairs(
-        filter.objectIds.map(objId => {
-          return [
-            objId,
-            opProjectArtifact({
-              project: projectNode,
-              artifactName: constString(objId),
-            }),
-          ];
-        })
-      ) as any
-    );
-  }
-
-  let artifactVersionsNode = opArray({} as any);
-  if (filter.latestOnly) {
-    artifactVersionsNode = opArtifactMembershipArtifactVersion({
-      artifactMembership: opArtifactLastMembership({
-        artifact: artifactsNode,
-      }),
-    }) as any;
-  } else {
-    artifactVersionsNode = opFlatten({
-      arr: opArtifactVersions({
-        artifact: artifactsNode,
-      }) as any,
-    });
-  }
-
-  // Filter to only Weave Objects
-  const weaveObjectsNode = opFilter({
-    arr: artifactVersionsNode,
-    filterFn: constFunction({row: 'artifactVersion'}, ({row}) => {
-      return opArtifactVersionIsWeaveObject({artifactVersion: row});
-    }),
-  });
-
-  // Build Keys
-  const dataNode = opMap({
-    arr: weaveObjectsNode,
-    mapFn: constFunction({row: 'artifactVersion'}, ({row}) => {
-      return opDict({
-        objectId: opArtifactName({
-          artifact: opArtifactVersionArtifactSequence({artifactVersion: row}),
-        }),
-        versionHash: opArtifactVersionHash({artifactVersion: row}),
-        dataDict: artifactVersionNodeToObjectVersionDictNode(row as any),
-      } as any);
-    }),
-  });
-
-  return dataNode;
-};
-
-export const useRootObjectVersions = (
-  entity: string,
-  project: string,
-  filter: ObjectVersionFilter,
-  opts?: {skip?: boolean}
-): Loadable<ObjectVersionSchema[]> => {
-  const dataNode = useRootObjectVersionsNode(entity, project, filter);
-  const dataValue = useNodeValue(dataNode, {skip: opts?.skip});
-
-  return useMemo(() => {
-    if (opts?.skip) {
-      return {
-        loading: false,
-        result: [],
-      };
-    }
-    const result = (dataValue.result ?? [])
-      .map((row: any) => ({
-        entity,
-        project,
-        objectId: row.objectId as string,
-        versionHash: row.versionHash as string,
-        path: 'obj',
-        refExtra: null,
-        versionIndex: row.dataDict.versionIndex as number,
-        typeName: row.dataDict.typeName as string,
-        category: typeNameToCategory(row.dataDict.typeName as string),
-        createdAtMs: row.dataDict.createdAtMs as number,
-      }))
-      .filter((row: any) => {
-        return (
-          filter.category == null || filter.category.includes(row.category)
-        );
-      })
-      // TODO: Move this to the weave filters?
-      .filter((row: any) => {
-        return !['OpDef', 'stream_table', 'type'].includes(row.typeName);
-      }) as ObjectVersionSchema[];
-
-    if (dataValue.loading) {
-      return {
-        loading: true,
-        result,
-      };
-    } else {
-      result.forEach(obj => {
-        setObjectVersionInCache(
-          {
-            entity,
-            project,
-            objectId: obj.objectId,
-            versionHash: obj.versionHash,
-            path: obj.path,
-            refExtra: obj.refExtra,
-          },
-          obj
-        );
-      });
-      return {
-        loading: false,
-        result,
-      };
-    }
-  }, [
-    dataValue.loading,
-    dataValue.result,
-    entity,
-    filter.category,
-    opts?.skip,
-    project,
-  ]);
-};
-
-type WFNaiveRefDict = {
-  entity: string;
-  project: string;
-  artifactName: string;
-  versionCommitHash: string;
-  filePathParts: string[];
-  refExtraTuples: Array<{
-    edgeType: string;
-    edgeName: string;
-  }>;
-};
-
-const refStringToRefDict = (uri: string): WFNaiveRefDict => {
-  const scheme = WANDB_ARTIFACT_REF_PREFIX;
-  if (!uri.startsWith(scheme)) {
-    throw new Error('Invalid uri: ' + uri);
-  }
-  const uriWithoutScheme = uri.slice(scheme.length);
-  let uriParts = uriWithoutScheme;
-  let refExtraPath = '';
-  const refExtraTuples = [];
-  if (uriWithoutScheme.includes('#')) {
-    [uriParts, refExtraPath] = uriWithoutScheme.split('#');
-    const refExtraParts = refExtraPath.split('/');
-    if (refExtraParts.length % 2 !== 0) {
-      throw new Error('Invalid uri: ' + uri);
-    }
-    for (let i = 0; i < refExtraParts.length; i += 2) {
-      refExtraTuples.push({
-        edgeType: refExtraParts[i],
-        edgeName: refExtraParts[i + 1],
-      });
-    }
-  }
-  const [entity, project, artifactNameAndVersion, filePath] = uriParts.split(
-    '/',
-    4
-  );
-  const [artifactName, versionCommitHash] = artifactNameAndVersion.split(':');
-  const filePathParts = filePath.split('/');
-
-  return {
-    entity,
-    project,
-    artifactName,
-    versionCommitHash,
-    filePathParts,
-    refExtraTuples,
+export type WFDataModelHooks = {
+  useCall: (key: CallKey | null) => Loadable<CallSchema | null>;
+  useCalls: (
+    entity: string,
+    project: string,
+    filter: CallFilter,
+    limit?: number,
+    opts?: {skip?: boolean}
+  ) => Loadable<CallSchema[]>;
+  useOpVersion: (key: OpVersionKey | null) => Loadable<OpVersionSchema | null>;
+  useOpVersions: (
+    entity: string,
+    project: string,
+    filter: OpVersionFilter,
+    limit?: number,
+    opts?: {skip?: boolean}
+  ) => Loadable<OpVersionSchema[]>;
+  useObjectVersion: (
+    key: ObjectVersionKey | null
+  ) => Loadable<ObjectVersionSchema | null>;
+  useRootObjectVersions: (
+    entity: string,
+    project: string,
+    filter: ObjectVersionFilter,
+    limit?: number,
+    opts?: {skip?: boolean}
+  ) => Loadable<ObjectVersionSchema[]>;
+  // Derived are under a subkey because they are not directly from the data model
+  // and the logic should be pushed into the core APIs. This is a temporary solution
+  // during the transition period.
+  derived: {
+    useChildCallsForCompare: (
+      entity: string,
+      project: string,
+      parentCallIds: string[],
+      selectedOpVersionRef: string | null,
+      selectedObjectVersionRef: string | null
+    ) => Loadable<
+      Array<{
+        parent: RawSpanFromStreamTableEra;
+        child: RawSpanFromStreamTableEra;
+      }>
+    >;
   };
-};
-
-/// Derived Interface - these are not part of the formal interface yet and need to be folded into the trace server directly ///
-
-export const useSubRunsFromWeaveQuery = (
-  entity: string,
-  project: string,
-  parentCallIds: string[] | undefined,
-  selectedOpVersionRef: string | null,
-  selectedObjectVersionRef: string | null
-): {
-  loading: boolean;
-  result: Array<{
-    parent: Span;
-    child: Span;
-  }>;
-} => {
-  const parentRunsNode = selectedObjectVersionRef
-    ? callsNode(entity, project, {
-        callIds: parentCallIds,
-        inputObjectVersionRefs: [selectedObjectVersionRef],
-      })
-    : opArray({} as any);
-  const childRunsNode = selectedOpVersionRef
-    ? callsNode(entity, project, {
-        opVersionRefs: [selectedOpVersionRef],
-      })
-    : opArray({} as any);
-  const joinedRuns = opJoin({
-    arr1: parentRunsNode,
-    arr2: childRunsNode,
-    join1Fn: constFunction({row: typedDict({span_id: 'string'})}, ({row}) => {
-      return opPick({obj: row, key: constString('span_id')});
-    }) as any,
-    join2Fn: constFunction({row: typedDict({parent_id: 'string'})}, ({row}) => {
-      return opPick({obj: row, key: constString('parent_id')});
-    }) as any,
-    alias1: constString('parent'),
-    alias2: constString('child'),
-    leftOuter: constBoolean(true),
-    rightOuter: constBoolean(false),
-  });
-  const nodeValue = useNodeValue(joinedRuns, {
-    skip: !selectedObjectVersionRef || !selectedOpVersionRef,
-  });
-  return useMemo(() => {
-    return {
-      loading: nodeValue.loading,
-      result: nodeValue.result ?? [],
-    };
-  }, [nodeValue.loading, nodeValue.result]);
-};
-
-//// In Mem Cache Layer ////
-
-const CACHE_SIZE = 5 * 2 ** 20; // 5MB
-
-const callCache = new LRUCache<string, CallSchema>({
-  max: CACHE_SIZE,
-  updateAgeOnGet: true,
-});
-
-const callCacheKeyFn = (key: CallKey) => {
-  return `call:${key.entity}/${key.project}/${key.callId}`;
-};
-
-const getCallFromCache = (key: CallKey) => {
-  return callCache.get(callCacheKeyFn(key));
-};
-
-const setCallInCache = (key: CallKey, value: CallSchema) => {
-  callCache.set(callCacheKeyFn(key), value);
-};
-
-const opVersionCache = new LRUCache<string, OpVersionSchema>({
-  max: CACHE_SIZE,
-  updateAgeOnGet: true,
-});
-
-const opVersionCacheKeyFn = (key: OpVersionKey) => {
-  return `op:${key.entity}/${key.project}/${key.opId}/${key.versionHash}`;
-};
-
-const getOpVersionFromCache = (key: OpVersionKey) => {
-  return opVersionCache.get(opVersionCacheKeyFn(key));
-};
-
-const setOpVersionInCache = (key: OpVersionKey, value: OpVersionSchema) => {
-  opVersionCache.set(opVersionCacheKeyFn(key), value);
-};
-
-const objectVersionCache = new LRUCache<string, ObjectVersionSchema>({
-  max: CACHE_SIZE,
-  updateAgeOnGet: true,
-});
-
-const objectVersionCacheKeyFn = (key: ObjectVersionKey) => {
-  return `obj:${key.entity}/${key.project}/${key.objectId}/${key.versionHash}/${key.path}/${key.refExtra}`;
-};
-
-const getObjectVersionFromCache = (key: ObjectVersionKey) => {
-  return objectVersionCache.get(objectVersionCacheKeyFn(key));
-};
-
-const setObjectVersionInCache = (
-  key: ObjectVersionKey,
-  value: ObjectVersionSchema
-) => {
-  objectVersionCache.set(objectVersionCacheKeyFn(key), value);
-};
-
-//// Utilities ////
-export const opVersionRefOpName = (opVersionRef: string) => {
-  return refUriToOpVersionKey(opVersionRef).opId;
-};
-
-// This one is a huge hack b/c it is based on the name. Once this
-// is added to the data model, we will need to make a query
-// wherever this is used!
-export const opVersionRefOpCategory = (opVersionRef: string) => {
-  return opNameToCategory(opVersionRefOpName(opVersionRef));
-};
-
-export const objectVersionNiceString = (ov: ObjectVersionSchema) => {
-  let result = ov.objectId;
-  if (ov.versionHash === '*') {
-    return result;
-  }
-  result += `:v${ov.versionIndex}`;
-  if (ov.path !== 'obj') {
-    result += `/${ov.path}`;
-  }
-  if (ov.refExtra) {
-    result += `#${ov.refExtra}`;
-  }
-  return result;
 };
